% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_incremental.R
\name{is_incremental}
\alias{is_incremental}
\alias{is_uniform}
\title{is_incremental}
\usage{
is_incremental(x, step = 1, na.action = NULL)

is_uniform(...)
}
\arguments{
\item{x}{numeric}

\item{step}{integer; step size for increment}

\item{na.action}{NULL or function to pre-process data}
}
\value{
logical
}
\description{
Test if a field is incrementally increasing or decreasing
}
\details{
\code{is_incremental} determines if x is incremental, i.e. ordered and equally
spaced.

\code{is_uniform} is a wrapper around \code{is_incremental} with \code{step=NULL}
}
\examples{

  is_incremental(1:5 )
  is_incremental( c(1,2,5))
 
}
\seealso{
\itemize{
\item \link{tests}
}
}
