% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sorted.R
\name{is_sorted}
\alias{is_sorted}
\title{is_sorted}
\usage{
is_sorted(x)
}
\arguments{
\item{x}{}

\item{decreasing}{logical; whether the sorting is de}
}
\description{
detemines whether a vector is sorted.
}
\details{
Returns \code{TRUE} if \code{x} is sorted.
}
\examples{

  is_sorted(1:10)
  is_sorted(-5:5)
  is_sorted(5:-5)
  
  is_sorted( letters )
  is_sorted( rev(letters) )
  
}
\seealso{
\itemize{
\item \code{\link[=is_increasing]{is_increasing()}}
\item \code{\link[base:is.unsorted]{base::is.unsorted()}}
}
}
