% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotone-package.R
\docType{package}
\name{monotone-package}
\alias{monotone-package}
\alias{monotone-package,}
\alias{monotone}
\title{Check/verify/confirm the monotonicity properties of a vector}
\description{
This packages provides a number of useful utilities for checking / verifying
or confirm the monotonicity properties of a vector. It can determine whether
values are constant, monotonic, or strictly increasing/decreasing.
It contains options for handling \code{NA} values as well as many
convenience and efficient functions.
}
\examples{
   montonocity(1:5) 
   montonocity(1:5) 
   
}
\author{
Christopher Brown
}
