% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{zero.action}
\alias{zero.action}
\alias{zero.omit}
\alias{zero.pass}
\alias{zero.replace}
\title{zero.action}
\usage{
zero.action(object, ...)

zero.pass(object, ...)

zero.omit(object, ...)

zero.replace(object, value, ...)
}
\arguments{
\item{object}{an R object, typically a numeric vector}
}
\description{
Action to take on zero.valued variables, similar to
\code{\link[stats]{na.action}}
}
\seealso{
\code{\link[stats]{na.action}}
}

