% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monotonicity.R
\name{monotonicity}
\alias{monotonicity}
\title{monotonicity}
\usage{
monotonicity(x, na.action = na.omit, zero.action = I, ...)
}
\arguments{
\item{x}{numeric vector}

\item{na.action}{function to handle \code{NA} values
(default:\code{\link[stats]{na.omit}})}

\item{...}{additional arguments passed to \code{na.action}}
}
\value{
integer;
\itemize{
\item \strong{2} : stictly increasing
\item \strong{1} : increasing / montonically increasing / non-decreasing
\item \strong{0} : not-monotone
\item \strong{-1}: decreasing / monotonically decreasing / non-increasing
\item \strong{-2}: strictly decreasing
\item \strong{\code{NA}}:\code{contains only}NA\code{s all}na.action\code{did not resolve all}NA`s
}
}
\description{
Access the monotonicity of a numeric vector
}
\details{
\code{monotonicty} determines the monotonicy (first derivative) of a numeric
vector as one of:
\itemize{
\item stictly increasing,
\item increasing / montonically increasing / non-decreasing,
\item decreasing / monotonically decreasing / non-increasing,
\item strictly decreasing, or
\item constant
}

Monotonicity tests, e.g. \link{is_increasing} are more efficient at testing.
}
\examples{
  monotonicity( 1:3 )      # 2  
  monotonicity( c(1,1,3) ) # 1 
  monotonicity( c(1,0,1) ) # 0  "No monotonicity, does not apply constant"
  monotonicity( c(3,1,1) ) # -1
  monotonicity( 3:1 )      # -2
  
}
\references{
http://en.wikipedia.org/wiki/Monotonic_function
http://stackoverflow.com/questions/13093912/how-to-check-if-a-sequence-of-numbers-is-monotonically-increasing-or-decreasing
}
