% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{monotonicity}
\alias{is.decreasing}
\alias{is.increasing}
\alias{monotonicity}
\title{monotonicity}
\usage{
monotonicity(x, na.action = na.omit, zero.action = I, ...)

is.increasing(x, type = "increasing", na.action = na.omit,
  zero.action = zero.pass)

is.decreasing(x, type = "increasing", na.action = na.omit,
  zero.action = zero.pass)
}
\arguments{
\item{x}{numeric vector}

\item{na.action}{function to handle \code{NA} values
(default:\code{\link[stats]{na.omit}})}

\item{...}{additional arguments passed to \code{na.action}

\code{monotonicty} determines the monotonicy of a numeric vector. Instead of
using diff, this function uses \code{cummax} and \code{cummin} functions for
efficiency. See the stackoverflow reference for timings and an explanation.}
}
\value{
integer;
     \code{2}  \tab for stictly increasing
     \code{1}  \tab increasing / montonically increasing / non-decreasing
     \code{0}  \tab constant
     \code{-1} \tab decreasing / monotonically decreasing / non-increasing
     \code{-2} \tab strictly decreasing
     \code{NA} \tab contains only \code{NAs} or \code(na.action) did not
     resolve all \code{NA}s
}
\description{
Determine the monotonicity of a numeric vector as:
  stictly increasing,
  increasing / montonically increasing / non-decreasing,
  decreasing / monotonically decreasing / non-increasing,
  strictly decreasing, or
  constant
}
\examples{
monotonicity( 1:3 )      # 2
  monotonicity( c(1,1,3) ) # 1
  monotonicity( c(1,0,1) ) # 0
  monotonicity( c(3,1,1) ) # -1
  monotonicity( 3:1 )      # -2

  is.constant(1:5)  # false
  is.constant(NA)
is.increasing( 1:10 )                  # TRUE
  is.increasing( c(1,2,1,3) )            # FALSE

  is.increasing( c(1,NA,2,3) )           # NA
  is.increasing( c(1,NA,2,3), na.omit )  # TRUE
}
\references{
http://en.wikipedia.org/wiki/Monotonic_function
  http://stackoverflow.com/questions/13093912/how-to-check-if-a-sequence-of-numbers-is-monotonically-increasing-or-decreasing
}

